Data is get from the sensors and afterwards once user saves them, it will save it to sharedPref and can be viewed later in history. Some bugs with navigation that has to be solved, was there before already as well. Decibel still not tested if working. DarkTheme now changes based if user has changed it in settings, bug with that is that you need to restart application for the user input to change theme. Notification+ShowInFahrenheit also saved now to sharedPref.
