+++ app/build.gradle
@@ -46,6 +46,7 @@ android {
 }
 
 dependencies {
+    implementation 'com.google.code.gson:gson:2.9.0'
     implementation "androidx.compose.runtime:runtime-livedata:$compose_version"
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
