+++ app/src/main/java/com/jonnesten/lullanap/SavedData.kt
@@ -1,0 +1,3 @@
+package com.jonnesten.lullanap
+
+data class SavedData(val date: String, val day: String, val lux: Float? = null, val temp: Float? = null, val noise: Float? = null, val review: Int? = null)
