+++ app/src/main/java/com/jonnesten/lullanap/SensorViewModel.kt
@@ -6,28 +6,30 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class SensorViewModel: ViewModel() {
-    private val _value: MutableLiveData<Float> = MutableLiveData()
-    private val _value2: MutableLiveData<Float> = MutableLiveData()
+    private val _temp: MutableLiveData<Float> = MutableLiveData()
+    private val _lux: MutableLiveData<Float> = MutableLiveData()
+    private val _dB: MutableLiveData<Float> = MutableLiveData()
     private val _clicked: MutableLiveData<Boolean> = MutableLiveData()
     private val _hasTempSensor: MutableLiveData<Boolean> = MutableLiveData()
     private val _hasLightSensor: MutableLiveData<Boolean> = MutableLiveData()
-    val value: LiveData<Float> = _value
-    val value2: LiveData<Float> = _value2
+    val temp: LiveData<Float> = _temp
+    val lux: LiveData<Float> = _lux
+    val dB: LiveData<Float> = _dB
     val clicked: LiveData<Boolean> = _clicked
     val hasTempSensor: LiveData<Boolean> = _hasTempSensor
     val hasLightSensor: LiveData<Boolean> = _hasLightSensor
 
 
-    fun updateValue(value: Float) {
-        _value.value = value
+    fun updateTemp(value: Float) {
+        _temp.value = value
     }
 
-    fun updateValue2(value: Float) {
-        _value2.value = value
+    fun updateLux(value: Float) {
+        _lux.value = value
     }
 
-    fun isClicked(value: Boolean) {
-        _clicked.value = value
+    fun updateDB(value: Float) {
+        _dB.value = value
     }
 
     fun updateTempSensor(value: Boolean) {
