+++ app/src/main/java/com/jonnesten/lullanap/ui/theme/Theme.kt
@@ -22,7 +22,8 @@ private val LightColorPalette = lightColors(
 
 
 @Composable
-fun LullaNapTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit) {
+fun LullaNapTheme(darkTheme: Boolean, content: @Composable () -> Unit) {
+
     val colors = if (darkTheme) {
         DarkColorPalette
     } else {
