+++ app/src/main/res/values-fi-rFI/strings.xml
@@ -20,4 +20,6 @@
     <string name="notifications">Ilmoitukset</string>
     <string name="show_in_fahrenheit">Näytä asteet Fahrenheiteissa</string>
     <string name="scanning">Skannataan…</string>
+    <string name="scan_again">Skannaa uudelleen</string>
+    <string name="save_data">Tallenna data</string>
 </resources>
