+++ app/src/main/res/values/strings.xml
@@ -19,4 +19,6 @@
     <string name="notifications">Notifications</string>
     <string name="show_in_fahrenheit">Show in Fahrenheit</string>
     <string name="scanning">Scanning…</string>
+    <string name="scan_again">Scan again</string>
+    <string name="save_data">Save data</string>
 </resources>
